@@ -47,8 +47,7 @@
   },
   "dependencies": {
     "URIjs": "1.16.0",
-    "brace": "0.5.1",
-    "d3": "3.5.5",
+    "brace": "0.8.0",
     "dagre": "0.7.2",
     "director": "1.2.8",
     "lodash": "3.9.3",
@@ -56,8 +55,9 @@
     "moment": "2.10.6",
     "normalize.css": "3.0.3",
     "object.observe": "0.2.4",
-    "react": "0.13.3",
-    "react-time": "4.0.0",
+    "react": "15.0.1",
+    "react-dom": "15.0.1",
+    "react-time": "4.0.2",
     "revalidator": "0.3.1",
     "st2client": "0.4.3"
   }
