@@ -1,135 +1,44 @@
 import _ from 'lodash';
-import d3 from 'd3';
+import React from 'react';
 import { EventEmitter } from 'events';
 
 import Arrow from './util/arrow';
 import bem from './util/bem';
-import { pack, unpack } from './util/packer';
 import Vector from './util/vector';
-
-import icons from './util/icon';
+import { pack, unpack } from './util/packer';
 
 const st2Class = bem('viewer')
-    , st2Icon = bem('icon')
     ;
 
-let nodeTmpl = (node) =>
-`
-  <div class="${st2Class('node-icon')}">
-    <img src="${icons.icons[node.pack] || ''}" width="32" height="32" />
-  </div>
-  <div class="${st2Class('node-content')}">
-    <form class="${st2Class('node-name-form')}">
-      <input type="text" class="${st2Class('node-name')}" value="${node.name}"/>
-    </form>
-    <div class="${st2Class('node-ref')}">${node.ref}</div>
-  </div>
-  <div class="${st2Class('node-actions')}">
-    <div class="${st2Class('node-edit')} ${st2Icon('edit')}"></div>
-    <div class="${st2Class('node-delete')} ${st2Icon('delete')}"></div>
-  </div>
-  <div class="${st2Class('node-buttons')}">
-    <span class="${st2Class('node-button')} ${st2Icon('success')}" draggable="true"></span>
-    <span class="${st2Class('node-button')} ${st2Icon('error')}" draggable="true"></span>
-    <span class="${st2Class('node-button')} ${st2Icon('complete')}" draggable="true"></span>
-  </div>
-`;
+import Node from './canvas/node';
+import Edge from './canvas/edge';
+import Label from './canvas/label';
 
-export default class Canvas extends EventEmitter {
-  constructor() {
-    super();
+let hiddenNode;
 
-    const self = this;
+class CanvasController extends EventEmitter {}
 
-    this.viewer = d3
-      .select(st2Class(null, true) + '')
-      ;
+export default class Canvas extends React.Component {
+  static propTypes = {
+    graph: React.PropTypes.object
+  }
 
-    const drag = d3.behavior.drag();
+  state = {
+    scale: 1
+  }
 
-    drag.on('dragstart', function () {
-      if (d3.event.sourceEvent.target === this) {
-        d3.select(this)
-          .classed(st2Class(null, 'grabbing'), true)
-          ;
-      }
-    });
+  nodes = {}
 
-    drag.on('dragend', function () {
-      if (d3.event.sourceEvent.target === this) {
-        d3.select(this)
-          .classed(st2Class(null, 'grabbing'), false)
-          ;
-      }
-    });
-
-    drag.on('drag', function () {
-      if (d3.event.sourceEvent.target === this) {
-        d3.event.sourceEvent.preventDefault();
-
-        const element = self.viewer.node();
-
-        element.scrollLeft -= d3.event.dx;
-        element.scrollTop -= d3.event.dy;
-      }
-    });
-
-    this.svg = this.viewer
-      .select(st2Class('canvas', true) + '')
-      .call(drag)
-      .on('dragover', function () {
-        if (d3.event.target === this) {
-          d3.event.stopPropagation();
-          self.dragOverOverlay(this, d3.event);
-        }
-      })
-      .on('dragenter', function () {
-        if (d3.event.target === this) {
-          d3.event.stopPropagation();
-          self.activateOverlay(this, d3.event);
-        }
-      })
-      .on('dragleave', function () {
-        if (d3.event.target === this) {
-          d3.event.stopPropagation();
-          self.deactivateOverlay(this, d3.event);
-        }
-      })
-      .on('drop', function () {
-        if (d3.event.target === this) {
-          d3.event.stopPropagation();
-          self.dropOnOverlay(this, d3.event);
-        }
-      });
-
-    const cStyle = window.getComputedStyle(this.svg.node());
-
-    this.paddings = {
-      top: parseInt(cStyle.getPropertyValue('padding-top')),
-      right: parseInt(cStyle.getPropertyValue('padding-right')),
-      bottom: parseInt(cStyle.getPropertyValue('padding-bottom')),
-      left: parseInt(cStyle.getPropertyValue('padding-left'))
-    };
-
-    this.clear();
-    this.resizeCanvas();
-
-    this.viewer
-      .attr('tabindex', '-1');
-
-    _.each(['keyup', 'keydown'], (type) => {
-      this.viewer
-        .on(type, () => {
-          this.emit(d3.event.type, d3.event);
-        })
-        ;
-    });
-
+  paddings = {
+    top: 45,
+    right: 10,
+    bottom: 10,
+    left: 10
   }
 
   toInner(x, y) {
-    x -= this.paddings.left;
-    y -= this.paddings.top;
+    x -= this.paddings.left / this.state.scale;
+    y -= this.paddings.top / this.state.scale;
 
     x = x < 0 ? 0 : x;
     y = y < 0 ? 0 : y;
@@ -138,375 +47,49 @@ export default class Canvas extends EventEmitter {
   }
 
   fromInner(x=0, y=0) {
-    x += this.paddings.left;
-    y += this.paddings.top;
+    if (_.isObject(x)) {
+      y = x.y;
+      x = x.x;
+    }
+
+    x += this.paddings.left / this.state.scale;
+    y += this.paddings.top / this.state.scale;
 
     return [x, y];
   }
 
-  clear() {
-    this.svg
-      .selectAll('g')
-      .remove('*');
+  componentDidMount() {
+    this._canvas = new CanvasController();
 
-    return this;
-  }
-
-  render(graph) {
-    this.graph = graph;
-
-    this.createNodes(this.viewer, this.graph);
-
-    this.reposition();
-  }
-
-  reposition() {
-    let nodes = this.viewer.selectAll(st2Class('node', true));
-
-    this.positionNodes(nodes, this.graph);
-    this.createEdgePaths(this.svg, this.graph);
-    this.createEdgeLabels(this.viewer, this.graph);
-  }
-
-  createNodes(selection, g) {
-    let self = this;
-
-    let nodes = selection
-      .selectAll(st2Class('node', true))
-      .data(g.nodes(), (v) => v)
-      ;
-
-    let enter = nodes.enter()
-      .append('div')
-        .attr('class', st2Class('node'))
-        .attr('draggable', 'true')
-        .html((d) => nodeTmpl(g.node(d)))
-        .on('click', function (name) {
-          d3.event.stopPropagation();
-          self.selectNode(this, d3.event, name);
-        })
-        .on('dragenter', function () {
-          d3.event.stopPropagation();
-          self.activateNode(this, d3.event);
-        })
-        .on('dragleave', function () {
-          d3.event.stopPropagation();
-          self.deactivateNode(this, d3.event);
-        })
-        .on('dragover', function () {
-          d3.event.stopPropagation();
-          self.dragOverNode(this, d3.event);
-        })
-        .on('dragstart', function (name) {
-          d3.event.stopPropagation();
-          d3.select(this)
-            .classed(st2Class('node', 'dragged'), true);
-          self.dragMove(this, d3.event, name);
-        })
-        .on('dragend', function () {
-          d3.event.stopPropagation();
-          d3.select(this)
-            .classed(st2Class('node', 'dragged'), false);
-        })
-        .on('drop', function (name) {
-          d3.event.stopPropagation();
-          self.dropOnNode(this, d3.event, name);
-        })
-        .each(d => {
-          const node = g.node(d);
-
-          node.on('change', (changes) => {
-            const refChanges = _.find(changes, {name: 'ref'});
-
-            if (refChanges) {
-              const target = d3.select(node.elem);
-
-              target
-                .select(st2Class('node-ref', true) + '')
-                .text(refChanges.object.ref);
-
-              target
-                .select(st2Class('node-icon', true) + '' + ' img')
-                .attr('src', icons.icons[node.pack] || '')
-                ;
-            }
-          });
-
-          g.on('select', (name) => {
-            d3.select(node.elem)
-              .classed(st2Class('node', 'selected'), name === d);
-          });
-        });
-
-    enter.select(st2Icon('success', true))
-      .on('dragstart', function (name) {
-        d3.event.stopPropagation();
-        self.dragSuccess(this, d3.event, name);
-      })
-      ;
-
-    enter.select(st2Icon('error', true))
-      .on('dragstart', function (name) {
-        d3.event.stopPropagation();
-        self.dragError(this, d3.event, name);
-      })
-      ;
-
-    enter.select(st2Icon('complete', true))
-      .on('dragstart', function (name) {
-        d3.event.stopPropagation();
-        self.dragComplete(this, d3.event, name);
-      })
-      ;
-
-    enter.select(st2Icon('delete', true))
-      .on('click', function (name) {
-        d3.event.stopPropagation();
-        self.deleteNode(this, d3.event, name);
-      })
-      ;
-
-    enter.select(st2Icon('edit', true))
-      .on('click', function (name) {
-        d3.event.stopPropagation();
-        self.edit(name);
-      })
-      ;
-
-    enter.select(st2Class('node-name', true) + '')
-      .on('keyup', () => d3.event.stopPropagation())
-      .on('keydown', () => d3.event.stopPropagation())
-      .on('blur', function (name) {
-        const value = this.value // HTMLInputElement
-            ;
-
-        if (value) {
-          self.emit('rename', name, value);
-        } else {
-          this.value = name;
-        }
-      })
-      ;
-
-    enter.select(st2Class('node-name-form', true) + '')
-      .on('submit', function () {
-        d3.event.preventDefault();
-
-        d3.select(this)
-          .select(st2Class('node-name', true) + '')
-            .node().blur()
-            ;
-      })
-      ;
-
-    nodes.exit()
-      .remove()
-      ;
-
-    nodes.each(function (name) {
-      let node = g.node(name)
-        , nodeElement = d3.select(this);
-
-      let {width, height} = nodeElement.node().getBoundingClientRect();
-
-      node.width = width;
-      node.height = height;
-
-      node.elem = this;
+    this._canvas.on('select', (selected) => {
+      this.setState({ selected });
     });
-
-    return nodes;
   }
 
-  positionNodes(selection, g) {
-    const transformer = (v) => {
-      let {x, y} = g.node(v);
-
-      [x, y] = this.fromInner(x, y);
-
-      return `translate(${x}px,${y}px)`;
-    };
-
-    selection.style('transform', transformer);
-    selection.style('-webkit-transform', transformer);
-  }
-
-  createEdgePaths(selection, g) {
-    this.resizeCanvas();
-
-    // Initialize selection with data set
-    let svgPaths = selection.selectAll(st2Class('edge', true))
-      .data(g.edges(), (e) => `${e.v}:${e.w}:${e.name}`);
-
-    const svgPathsEnter = svgPaths.enter()
-      .append('g')
-        .attr('class', (e) => st2Class('edge') + ' ' + st2Class('edge', g.edge(e).type))
-        ;
-
-    svgPathsEnter.append('path')
-      .attr('class', st2Class('edge-path'))
-      ;
-
-    svgPathsEnter.append('defs');
-
-    const svgPathExit = svgPaths.exit();
-
-    svgPathExit
-      .remove();
-
-    svgPaths
-      .each(function(e) {
-        const element = d3.select(this)
-            , edge = g.edge(e)
-            ;
-
-        edge.arrowheadId = _.uniqueId('arrowhead');
-
-        const tail = g.node(e.v)
-            , head = g.node(e.w)
-            , A = tail.intersect(head)
-            , B = head.intersect(tail)
-            // Shift line back a little
-            , AB = B.subtract(A)
-            , delta = AB.unit().multiply(Arrow.size.x)
-            , points = [A.subtract(delta), B.subtract(delta)]
-            ;
-
-        const line = d3.svg.line()
-          .x((d) => d.x)
-          .y((d) => d.y)
-          ;
-
-        element.select(st2Class('edge-path', true) + '')
-          .attr('marker-end', () => `url(#${edge.arrowheadId})`)
-          .style('fill', 'none')
-          .attr('d', line(points))
-          ;
-      });
-
-    // Add arrow shape
-    svgPaths.selectAll('defs *').remove();
-    svgPaths.selectAll('defs')
-      .each(function(e) {
-        let edge = g.edge(e)
-          ;
-        new Arrow(d3.select(this), edge.arrowheadId, edge, 'arrowhead');
-      });
-  }
-
-  createEdgeLabels(selection, g) {
-    const self = this
-        , labels = selection
-            .selectAll(st2Class('label', true))
-            .data(g.edges(), (e) => `${e.v}:${e.w}:${e.name}`)
-            ;
-
-    labels.enter()
-      .append('div')
-        .attr('class', (e) => st2Class('label') + ' ' + st2Class('label', g.edge(e).type))
-        .on('click', function (edge) {
-          d3.event.stopPropagation();
-          self.deleteEdge(this, d3.event, edge);
-        })
-        ;
-
-    labels.exit()
-      .remove()
-      ;
-
-    const transformer = (e) => {
-      const head = g.node(e.v)
-          , tail = g.node(e.w)
-          , [A, B] = [head.intersect(tail), tail.intersect(head)]
-          // find mid point on the line excluding arrow
-          , AB = B.subtract(A)
-          , length = AB.length() - Arrow.size.x
-          , C = AB.unit().multiply(length/2).add(A)
-          , [x, y] = this.fromInner(C.x, C.y)
-          ;
-
-      return `translate(${x}px,${y}px)`;
-    };
-
-    labels
-      .style('transform', transformer)
-      .style('-webkit-transform', transformer)
-      ;
-  }
-
-  centerElement() {
-    if (this.element) {
-      let canvasBounds = this.svg[0][0].getBoundingClientRect()
-        , elementBounds = this.element[0][0].getBoundingClientRect();
-
-      let xCenterOffset = (canvasBounds.width - elementBounds.width) / 2;
-      let yCenterOffset = (canvasBounds.height - elementBounds.height) / 2;
-
-      const transformer = 'translate(' + xCenterOffset + ', ' + yCenterOffset + ')';
-
-      this.element.attr('transform', transformer);
-    }
-  }
-
-  resizeCanvas() {
-    let element = this.viewer.node()
-      , dimensions = {
-        width: element.offsetWidth,
-        height: element.offsetHeight
-      };
-
-    if (this.graph) {
-       dimensions = _.reduce(this.graph.nodes(), (acc, name) => {
-        let {x, y, width, height} = this.graph.node(name);
-
-        [x, y] = this.fromInner(x, y);
-
-        x += width + this.paddings.right;
-        y += height + this.paddings.bottom;
-
-        acc.width = acc.width < x ? x : acc.width;
-        acc.height = acc.height < y ? y : acc.height;
-
-        return acc;
-      }, dimensions);
-    }
-
-    element.style.setProperty('overflow', 'hidden');
-    element.offsetHeight; // trigger a recalc
-    element.style.setProperty('overflow', 'auto');
-
-    this.svg.attr('width', dimensions.width);
-    this.svg.attr('height', dimensions.height);
-  }
-
-  focus() {
-    this.viewer
-      .node()
-        .focus();
+  on(...args) {
+    return this._canvas.on(...args);
   }
 
   edit(name) {
-    const node = this.graph.node(name);
-    d3.select(node.elem)
-      .classed(st2Class('node', 'edited'), true)
-      .select(st2Class('node-name', true) + '')
-        .node().select() // This one is HTMLInputElement.select, not d3.select
-        ;
+    return this.nodes[name].edit(name);
+  }
+
+  focus() {
+    return this.refs.viewer.focus();
   }
 
   show(name) {
-    const node = this.graph.node(name)
-        , view = this.viewer.node()
-        , { x, y } = node
-        , { scrollWidth: width, scrollHeight: height } = node.elem
+    const node = this._canvas.graph.node(name)
+        , view = this.refs.viewer
+        , { x, y, width, height } = node
         , { scrollLeft, scrollTop, clientWidth, clientHeight } = view
-        , [ viewWidth, viewHeight ] = this.toInner(clientWidth, clientHeight)
+        , [ viewWidth, viewHeight ] = [clientWidth, clientHeight]//this.toInner(clientWidth, clientHeight)
         ;
 
     const A = new Vector(scrollLeft, scrollTop)
         , B = new Vector(x, y)
         , C = new Vector(x + width, y + height)
-        , D = new Vector(...this.toInner(scrollLeft + viewWidth, scrollTop + viewHeight))
+        , D = new Vector(...[scrollLeft + viewWidth, scrollTop + viewHeight])//this.toInner(scrollLeft + viewWidth, scrollTop + viewHeight))
         , AB = B.subtract(A)
         , CD = D.subtract(C)
         ;
@@ -528,155 +111,330 @@ export default class Canvas extends EventEmitter {
     }
   }
 
-  // Event Handlers
+  draw(graph) {
+    const nodes = graph._nodes;
+    const edges = _.mapValues(graph._edgeObjs, (edge, key) => {
+      const { v, w } = edge;
 
-  activateOverlay(element) {
-    element.classList.add(st2Class('canvas', 'active'));
+      return Object.assign({
+        arrowheadId: _.uniqueId('arrowhead'),
+        v: graph._nodes[v],
+        w: graph._nodes[w]
+      }, graph._edgeLabels[key]);
+    });
+
+    this.setState({ nodes, edges });
+    this._canvas.graph = graph;
   }
 
-  deactivateOverlay(element) {
-    element.classList.remove(st2Class('canvas', 'active'));
+  handleLabelClick(e, edge) {
+    const { v, w } = edge;
+    this._canvas.emit('disconnect', { v: v.name, w: w.name });
   }
 
-  dragOverOverlay(element, event) {
-    let dt = event.dataTransfer;
-
-    if (_.includes(['move', 'copy', 'all'], dt.effectAllowed)) {
-      event.preventDefault();
-    }
+  handleNodeSelect(e, name) {
+    this._canvas.emit('select', name, e);
+    // this.setState({ selected });
   }
 
-  dropOnOverlay(element, event) {
-    let packet;
+  handleNodePick(e, name) {
+    const dt = e.dataTransfer;
 
-    packet = event.dataTransfer.getData('nodePack');
-    if (packet) {
-      let { name, offsetX, offsetY } = unpack(packet)
-        , {offsetX: x, offsetY: y} = event // Relative to itself (Viewer)
-        ;
+    const {layerX: targetX, layerY: targetY} = e.nativeEvent;
+    const { x, y } = this.state.nodes[name];
 
-      [x, y] = this.toInner(x - offsetX, y - offsetY);
+    const vViewport = new Vector(targetX, targetY);
+    const vNode = new Vector(x, y);
+    const vOrigin = new Vector(this.paddings.left, this.paddings.top);
 
-      this.emit('move', name, x, y);
-      this.deactivateOverlay(element);
-      return;
-    }
+    const offset = vViewport.subtract(vOrigin).divide(this.state.scale).subtract(vNode);
 
-    packet = event.dataTransfer.getData('actionPack');
-    if (packet) {
-      let { action } = unpack(packet)
-        , {offsetX: x, offsetY: y} = event // Relative to itself (Viewer)
-        ;
-
-      [x, y] = this.toInner(x, y);
-
-      this.emit('create', action, x, y);
-      this.deactivateOverlay(element);
-      return;
-    }
-  }
-
-  selectNode(element, event, name) {
-    this.emit('select', name, event);
-  }
-
-  activateNode(element) {
-    element.classList.add(st2Class('node', 'active'));
-  }
-
-  deactivateNode(element) {
-    element.classList.remove(st2Class('node', 'active'));
-  }
-
-  dragOverNode(element, event) {
-    let dt = event.dataTransfer;
-
-    if (_.includes(['link', 'all'], dt.effectAllowed)) {
-      event.preventDefault();
-    }
-  }
-
-  dropOnNode(element, event, name) {
-    event.stopPropagation();
-
-    let dt = event.dataTransfer
-      , {source, type} = unpack(dt.getData('linkPack'))
-      , destination = name
-      ;
-
-    this.emit('link', source, destination, type);
-    this.deactivateNode(element);
-  }
-
-  dragMove(element, event, name) {
-    let dt = event.dataTransfer
-      , {layerX: x, layerY: y} = event // Relative to the closest positioned element (Viewer)
-      , node = this.graph.node(name)
-      , [offsetX, offsetY] = this.toInner(x - node.x, y - node.y)// [x - node.x, y - node.y]
-      ;
-
-    const crt = node.elem.cloneNode(true);
+    const crt = e.target.cloneNode(true);
     crt.style.removeProperty('transform');
     crt.style.removeProperty('-webkit-transform');
     crt.style.setProperty('z-index', -1);
 
-    if (this._hiddenNode) {
-      this._hiddenNode.parentNode.replaceChild(crt, this._hiddenNode);
+    if (hiddenNode) {
+      hiddenNode.parentNode.replaceChild(crt, hiddenNode);
     } else {
       document.body.appendChild(crt);
     }
 
-    this._hiddenNode = crt;
+    hiddenNode = crt;
 
-    dt.setDragImage(crt, offsetX, offsetY);
-    dt.setData('nodePack', pack({ name, offsetX, offsetY }));
+    dt.setDragImage(crt, offset.x, offset.y);
+    dt.setData('nodePack', pack({ name, offsetX: offset.x, offsetY: offset.y }));
     dt.effectAllowed = 'move';
   }
 
-  dragSuccess(element, event, name) {
-    let dt = event.dataTransfer;
-
-    dt.setData('linkPack', pack({
-      source: name,
-      type: 'success'
-    }));
-    dt.effectAllowed = 'link';
+  handleNodeRename(e, name, value) {
+    this._canvas.emit('rename', name, value);
   }
 
-  dragError(element, event, name) {
-    let dt = event.dataTransfer;
-
-    dt.setData('linkPack', pack({
-      source: name,
-      type: 'error'
-    }));
-    dt.effectAllowed = 'link';
+  handleNodeDelete(e, name) {
+    this._canvas.emit('delete', name);
   }
 
-  dragComplete(element, event, name) {
-    let dt = event.dataTransfer;
-
-    dt.setData('linkPack', pack({
-      source: name,
-      type: 'complete'
-    }));
-    dt.effectAllowed = 'link';
+  handleNodeConnect(e, source, destination, type) {
+    this._canvas.emit('link', source, destination, type);
   }
 
-  dragDisconnect(element, event, name) {
-    let dt = event.dataTransfer;
-
-    dt.setData('linkPack', pack({
-      source: name
-    }));
-    dt.effectAllowed = 'link';
+  handleCanvasDragEnter(e) {
+    e.stopPropagation();
+    this.setState({ active: true });
   }
 
-  deleteNode(element, event, name) {
-    this.emit('delete', name);
+  handleCanvasDragLeave(e) {
+    e.stopPropagation();
+    this.setState({ active: false });
   }
 
-  deleteEdge(element, event, edge) {
-    this.emit('disconnect', edge);
+  handleCanvasDragOver(e) {
+    e.stopPropagation();
+    const dt = e.dataTransfer;
+
+    if (_.includes(['move', 'copy', 'all'], dt.effectAllowed)) {
+      e.preventDefault();
+    }
+  }
+
+  handleCanvasDrop(e) {
+    e.stopPropagation();
+
+    let packet;
+
+    packet = e.dataTransfer.getData('nodePack');
+    if (packet) {
+      let { name, offsetX, offsetY } = unpack(packet)
+        , {offsetX: x, offsetY: y} = e.nativeEvent // Relative to itself (Viewer)
+        ;
+
+      [x, y] = [x - offsetX, y - offsetY]; //this.toInner(x - offsetX, y - offsetY);
+
+      this._canvas.emit('move', name, x, y);
+      this.setState({ active: false });
+      return;
+    }
+
+    packet = e.dataTransfer.getData('actionPack');
+    if (packet) {
+      let { action } = unpack(packet)
+        , {offsetX: x, offsetY: y} = e.nativeEvent // Relative to itself (Viewer)
+        ;
+
+      // [x, y] = this.toInner(x, y);
+
+      this._canvas.emit('create', action, x, y);
+      this.setState({ active: false });
+      return;
+    }
+  }
+
+  handleCanvasMouseMove(e) {
+    if (this.state.grabbing) {
+      e.stopPropagation();
+      e.preventDefault();
+
+      const element = this.refs.viewer;
+
+      const { clientX, clientY } = e.nativeEvent;
+
+      const A1 = this.state.grabbing
+        , B = new Vector(clientX, clientY)
+        , { x, y } = A1.subtract(B)
+        ;
+
+      element.scrollLeft = x;
+      element.scrollTop = y;
+    }
+  }
+
+  handleCanvasMouseDown(e) {
+    e.preventDefault(); // prevents text getting selected during mouse drag
+    e.stopPropagation();
+
+    const { scrollLeft, scrollTop } = this.refs.viewer;
+    const { clientX, clientY } = e.nativeEvent;
+
+    const P = new Vector(scrollLeft, scrollTop)
+        , A = new Vector(clientX, clientY)
+        ;
+
+    this.setState({ grabbing: A.add(P) });
+  }
+
+  handleCanvasMouseUp(e) {
+    e.preventDefault();
+    e.stopPropagation();
+
+    this.setState({ grabbing: false });
+  }
+
+  handleCanvasMouseEnter(e) {
+    if (this.state.grabbing && e.nativeEvent.buttons !== 1) {
+      this.setState({ grabbing: false });
+    }
+  }
+
+  handleCanvasOnWheel(e) {
+    e.preventDefault();
+
+    const delta = e.deltaY;
+    let { scale } = this.state;
+
+    scale += parseInt(delta) / 1000;
+
+    if (scale < .1) {
+      scale = .1;
+    }
+
+    if (scale > 1) {
+      scale = 1;
+    }
+
+    this.setState({ scale });
+  }
+
+  getSvgSize() {
+    const { nodes={}, scale } = this.state;
+
+    let { offsetWidth: width, offsetHeight: height } = this.refs.viewer;
+
+    const { top, bottom, left, right } = this.paddings;
+
+    width -= left + right;
+    height -= top + bottom;
+
+    width /= scale;
+    height /= scale;
+
+    for (const key of Object.keys(nodes)) {
+      const { width: w, height: h } = nodes[key];
+      let [ x, y ] = this.fromInner(nodes[key]);
+
+      x += w;
+      y += h;
+
+      width = width < x ? x : width;
+      height = height < y ? y : height;
+    }
+
+    return { width, height };
+  }
+
+  render() {
+    const containerProps = {
+      style: {
+        height: '100%'
+      },
+      onMouseMove: (e) => this.handleCanvasMouseMove(e),
+      onMouseDown: (e) => this.handleCanvasMouseDown(e),
+      onMouseUp: (e) => this.handleCanvasMouseUp(e),
+      onMouseEnter: (e) => this.handleCanvasMouseEnter(e),
+      onWheel: (e) => this.handleCanvasOnWheel(e)
+    };
+
+    const viewerProps = {
+      className: st2Class(null),
+      tabindex: -1,
+      ref: 'viewer'
+    };
+
+    if (this.state.active) {
+      viewerProps.className += ' ' + st2Class(null, 'active');
+    }
+
+    if (this.state.grabbing) {
+      viewerProps.className += ' ' + st2Class(null, 'grabbing');
+    }
+
+    const zoomerProps = {
+      className: st2Class('zoomer'),
+      style: {
+        transform: `scale(${this.state.scale})`
+      }
+    };
+
+    const { top, bottom, left, right } = this.paddings;
+
+    const canvasProps = {
+      className: st2Class('canvas'),
+      style: {
+        margin: [top, right, bottom, left].map(v => (v / this.state.scale) + 'px').join(' ')
+      },
+      onDragEnter: (e) => this.handleCanvasDragEnter(e),
+      onDragLeave: (e) => this.handleCanvasDragLeave(e),
+      onDragOver: (e) => this.handleCanvasDragOver(e),
+      onDrop: (e) => this.handleCanvasDrop(e)
+    };
+
+    if (this.refs.viewer) {
+      const { width, height } = this.getSvgSize();
+      Object.assign(canvasProps, { width, height });
+    }
+
+    return <div {...containerProps} >
+      <div {...viewerProps} >
+        <div {...zoomerProps} >
+          <svg {...canvasProps} >
+            {
+              _.map(this.state.edges, (edge, key) => {
+                const props = {
+                  key,
+                  value: edge
+                };
+
+                return <Edge {...props} />;
+              })
+            }
+          </svg>
+          {
+            _.map(this.state.edges, (edge, key) => {
+              const { v, w } = edge;
+
+              const A = w.intersect(v)
+                  , B = v.intersect(w)
+                  // find mid point on the line excluding arrow
+                  , AB = B.subtract(A)
+                  , length = AB.length() + Arrow.size.x
+                  , M = AB.unit().multiply(length/2).add(A)
+                  , [x, y] = this.fromInner(M)
+                  ;
+
+              const props = {
+                key,
+                value: edge,
+                x,
+                y,
+                onClick: (e) => this.handleLabelClick(e, edge)
+              };
+
+              return <Label {...props} />;
+            })
+          }
+          {
+            _.map(this.state.nodes, (node, key) => {
+              const [ x, y ] = this.fromInner(node);
+
+              const props = {
+                key,
+                value: node,
+                x,
+                y,
+                selected: key === this.state.selected,
+                ref: (c) => this.nodes[key] = c,
+                onSelect: (e) => this.handleNodeSelect(e, key),
+                onPick: (e) => this.handleNodePick(e, key),
+                onRename: (e, v) => this.handleNodeRename(e, key, v),
+                onDelete: (e) => this.handleNodeDelete(e, key),
+                onConnect: (...args) => this.handleNodeConnect(...args)
+              };
+
+              return <Node {...props} />;
+            })
+          }
+        </div>
+      </div>
+    </div>;
   }
 }
